던전 재시작 관련 문제들 수정

+++ Assets/Scripts/Manager/PartyManager.cs
@@ -105,7 +105,7 @@ public class PartyManager : MonoBehaviour
         return null;
     }
 
-    public void AddExp(int exp)
+    public void AddExp(float exp)
     {
         foreach(Hero hero in heroParty)
         {
