+++ Assets/Scripts/UI/PlayerInfoUI.cs
@@ -14,11 +14,11 @@ public class PlayerInfoUI : MonoBehaviour
     public Text maxHero;
 
     // UI 갱신
-    public void RenewLV(int lv)             { LV.text = lv.ToString(); }
-    public void RenewExp(int exp, int max)  { Exp.fillAmount = (float)exp / max * 0.75f; }
-    public void RenewGold(int value)        { gold.text = value.ToString(); }
-    public void RenewCurParty(int value)    { curParty.text = value.ToString(); }
-    public void RenewMaxParty(int value)    { maxParty.text = value.ToString(); }
-    public void RenewCurHero(int value)    { curHero.text = value.ToString(); }
-    public void RenewMaxHero(int value)    { maxHero.text = value.ToString(); }
+    public void RenewLV(int lv)                 { LV.text = lv.ToString(); }
+    public void RenewExp(float exp, float max)  { Exp.fillAmount = exp / max * 0.75f; }
+    public void RenewGold(int value)            { gold.text = value.ToString(); }
+    public void RenewCurParty(int value)        { curParty.text = value.ToString(); }
+    public void RenewMaxParty(int value)        { maxParty.text = value.ToString(); }
+    public void RenewCurHero(int value)         { curHero.text = value.ToString(); }
+    public void RenewMaxHero(int value)         { maxHero.text = value.ToString(); }
 }
